+++ src/nanonanoda.rs
@@ -20,7 +20,7 @@ use soundlog::{VgmBuilder, VgmDocument};
 #[derive(Debug, Clone)]
 pub struct SpectralFeature {
     pub fnumber: FNumber,
-    pub magnitude: f64,
+    pub magnitude: f32,
 }
 
 /// Analyze a mono sample window and map dominant spectral peaks to
@@ -268,7 +268,7 @@ pub fn process_samples_resynth_multi(
 
 // helper: map FFT magnitude to TL (0 = loud, larger value = quieter)
 // `max_tl` is the maximum TL value to use (e.g. 0x24). note: 0x00 == loudest.
-pub fn mag_to_tl(mag: f64, max_tl: u8) -> u8 {
+pub fn mag_to_tl(mag: f32, max_tl: u8) -> u8 {
     // For non-finite or silence, return full attenuation (0x3f)
     if !mag.is_finite() || mag <= 0.0 {
         return 0x3f;
@@ -278,9 +278,9 @@ pub fn mag_to_tl(mag: f64, max_tl: u8) -> u8 {
     let db_max = 0.0;
     let t = ((mag_db - db_min) / (db_max - db_min)).clamp(0.0, 1.0);
     // Map t==1.0 -> max_tl (loud), t==0.0 -> 0x3f (silent)
-    let range = (0x3f_i32 - max_tl as i32) as f64;
-    let tl_f = (1.0 - t) * range + (max_tl as f64);
-    tl_f.round().clamp(max_tl as f64, 0x3f as f64) as u8
+    let range = (0x3f_i32 - max_tl as i32) as f32;
+    let tl_f = (1.0 - t) * range + (max_tl as f32);
+    tl_f.round().clamp(max_tl as f32, 0x3f as f32) as u8
 }
 
 /// Similar to `process_samples_resynth_multi`, but instead of synthesizing
